LPS-69238 Move everything that is not exported to internal package
