+++ adaptive-media-image-jax-rs/src/main/java/com/liferay/adaptive/media/image/jaxrs/internal/ImageAdaptiveMediaRootResource.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.adaptive.media.image.jaxrs;
+package com.liferay.adaptive.media.image.jaxrs.internal;
 
 import com.liferay.adaptive.media.image.configuration.ImageAdaptiveMediaConfigurationHelper;
