+++ adaptive-media-image-jax-rs/src/main/resources/configuration/com.liferay.portal.remote.rest.extender.configuration.RestExtenderConfiguration-adaptivemediajaxrs
@@ -1,2 +1,2 @@
 contextPaths=/adaptive-media
-jaxrs.application.filters=(component.name=com.liferay.adaptive.media.image.jaxrs.ImageAdaptiveMediaJaxRApplication)
+jaxrs.application.filters=(component.name=com.liferay.adaptive.media.image.jaxrs.internal.ImageAdaptiveMediaJaxRApplication)
